Fix build errors via file-renaming and reinstalling packages
