+++ frontend/pages/viewer-registration/index.tsx
