+++ frontend/tests/snapshots/dashboard.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import Dashboard from "../../pages/dashboard/dashboard"
+import Dashboard from "../../pages/dashboard"
 import renderer from "react-test-renderer"
 
 it("renders Dashboard correctly", () => {
